refactor: 이미지 업로드 관련 컴포넌트 리팩토링 - 훅으로 S3 관련 함수를 사용하도록 변경 - Props drilling이 없도록 중간에 불필요한 컴포넌트 삭제
